LayoutTests/canvas/philip/tests/2d.imageData.get.order.rgb-expected.txt

< [index] >
2d.imageData.get.order.rgb
getImageData() returns R then G then B
References: 2d.pixelarray.order 2d.pixelarray.indexes
Actual output:
Passed
